requirements.txt
Flask==2.3.2
tensorflow==2.13.0
numpy
Pillow
requests
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
supabase
werkzeug
